problem-solver/py/requirements.txt
py-sc-kpm==0.2.0
py-sc-client==0.3.0
requests==2.31.0
python-telegram-bot==21.6
